[Change] Move SavedController to controller dic
